requirements.txt
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
celery[redis]==5.2.7
flower==1.2.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
requests==2.28.1
boto3==1.26.56
botocore==1.29.56